requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5
requests==2.28.2
pydantic==1.10.13
streamlit==1.23.1
